LPS-124961 Move class ending with *Constants to package *.constants
