+++ modules/apps/commerce/commerce-account-admin-web/src/main/java/com/liferay/commerce/account/admin/web/internal/servlet/taglib/ui/CommerceAccountDetailScreenNavigationEntry.java
@@ -14,6 +14,7 @@
 
 package com.liferay.commerce.account.admin.web.internal.servlet.taglib.ui;
 
+import com.liferay.commerce.account.admin.web.internal.servlet.taglib.ui.constants.CommerceAccountScreenNavigationConstants;
 import com.liferay.commerce.account.model.CommerceAccount;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
