+++ modules/apps/commerce/commerce-account-admin-web/src/main/java/com/liferay/commerce/account/admin/web/internal/servlet/taglib/ui/CommerceAccountOrganizationRelScreenNavigationEntry.java
@@ -15,6 +15,7 @@
 package com.liferay.commerce.account.admin.web.internal.servlet.taglib.ui;
 
 import com.liferay.commerce.account.admin.web.internal.display.context.CommerceAccountOrganizationRelAdminDisplayContext;
+import com.liferay.commerce.account.admin.web.internal.servlet.taglib.ui.constants.CommerceAccountScreenNavigationConstants;
 import com.liferay.commerce.account.model.CommerceAccount;
 import com.liferay.commerce.account.service.CommerceAccountOrganizationRelService;
 import com.liferay.commerce.account.service.CommerceAccountService;
