+++ modules/apps/commerce/commerce-account-admin-web/src/main/java/com/liferay/commerce/account/admin/web/internal/servlet/taglib/ui/CommerceAccountUserRelScreenNavigationEntry.java
@@ -15,6 +15,7 @@
 package com.liferay.commerce.account.admin.web.internal.servlet.taglib.ui;
 
 import com.liferay.commerce.account.admin.web.internal.display.context.CommerceAccountUserRelAdminDisplayContext;
+import com.liferay.commerce.account.admin.web.internal.servlet.taglib.ui.constants.CommerceAccountScreenNavigationConstants;
 import com.liferay.commerce.account.model.CommerceAccount;
 import com.liferay.commerce.account.service.CommerceAccountService;
 import com.liferay.commerce.account.service.CommerceAccountUserRelService;
