+++ modules/apps/commerce/commerce-account-admin-web/src/main/java/com/liferay/commerce/account/admin/web/internal/servlet/taglib/ui/constants/CommerceAccountScreenNavigationConstants.java
@@ -1,0 +1,37 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.commerce.account.admin.web.internal.servlet.taglib.ui.constants;
+
+/**
+ * @author Alessio Antonio Rendina
+ */
+public class CommerceAccountScreenNavigationConstants {
+
+	public static final String CATEGORY_KEY_COMMERCE_ACCOUNT_DETAIL = "detail";
+
+	public static final String ENTRY_KEY_COMMERCE_ACCOUNT_ADDRESSES =
+		"addresses";
+
+	public static final String ENTRY_KEY_COMMERCE_ACCOUNT_DETAIL = "detail";
+
+	public static final String ENTRY_KEY_COMMERCE_ACCOUNT_ORGANIZATIONS =
+		"organizations";
+
+	public static final String ENTRY_KEY_COMMERCE_ACCOUNT_USERS = "users";
+
+	public static final String SCREEN_NAVIGATION_KEY_COMMERCE_ACCOUNT_GENERAL =
+		"commerce.account.general";
+
+}
