+++ modules/apps/commerce/commerce-account-admin-web/src/main/resources/META-INF/resources/init.jsp
@@ -29,7 +29,7 @@ taglib uri="http://liferay.com/tld/ui" prefix="liferay-ui" %>
 page import="com.liferay.commerce.account.admin.web.internal.display.context.CommerceAccountAdminDisplayContext" %><%@
 page import="com.liferay.commerce.account.admin.web.internal.display.context.CommerceAccountOrganizationRelAdminDisplayContext" %><%@
 page import="com.liferay.commerce.account.admin.web.internal.display.context.CommerceAccountUserRelAdminDisplayContext" %><%@
-page import="com.liferay.commerce.account.admin.web.internal.servlet.taglib.ui.CommerceAccountScreenNavigationConstants" %><%@
+page import="com.liferay.commerce.account.admin.web.internal.servlet.taglib.ui.constants.CommerceAccountScreenNavigationConstants" %><%@
 page import="com.liferay.commerce.account.constants.CommerceAccountActionKeys" %><%@
 page import="com.liferay.commerce.account.constants.CommerceAccountConstants" %><%@
 page import="com.liferay.commerce.account.exception.CommerceAccountNameException" %><%@
