+++ modules/apps/commerce/commerce-account-group-admin-web/src/main/java/com/liferay/commerce/account/group/admin/web/internal/servlet/taglib/ui/CommerceAccountGroupAccountScreenNavigationEntry.java
@@ -14,6 +14,7 @@
 
 package com.liferay.commerce.account.group.admin.web.internal.servlet.taglib.ui;
 
+import com.liferay.commerce.account.group.admin.web.internal.servlet.taglib.ui.constants.CommerceAccountGroupScreenNavigationConstants;
 import com.liferay.commerce.account.model.CommerceAccountGroup;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
 import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
