+++ modules/apps/commerce/commerce-account-group-admin-web/src/main/resources/META-INF/resources/init.jsp
@@ -30,7 +30,7 @@ page import="com.liferay.commerce.account.exception.CommerceAccountGroupNameExce
 page import="com.liferay.commerce.account.exception.NoSuchAccountGroupCommerceAccountRelException" %><%@
 page import="com.liferay.commerce.account.exception.NoSuchAccountGroupException" %><%@
 page import="com.liferay.commerce.account.group.admin.web.internal.display.context.CommerceAccountGroupAdminDisplayContext" %><%@
-page import="com.liferay.commerce.account.group.admin.web.internal.servlet.taglib.ui.CommerceAccountGroupScreenNavigationConstants" %><%@
+page import="com.liferay.commerce.account.group.admin.web.internal.servlet.taglib.ui.constants.CommerceAccountGroupScreenNavigationConstants" %><%@
 page import="com.liferay.commerce.account.model.CommerceAccount" %><%@
 page import="com.liferay.commerce.account.model.CommerceAccountGroup" %><%@
 page import="com.liferay.commerce.account.model.CommerceAccountGroupCommerceAccountRel" %><%@
