+++ modules/apps/commerce/commerce-account-web/src/main/resources/META-INF/resources/init.jsp
@@ -37,7 +37,7 @@ page import="com.liferay.commerce.account.web.internal.frontend.CommerceAccountC
 page import="com.liferay.commerce.account.web.internal.frontend.CommerceAccountOrganizationClayDataSetDataSetDisplayView" %><%@
 page import="com.liferay.commerce.account.web.internal.frontend.CommerceAccountUserClayDataSetDataSetDisplayView" %><%@
 page import="com.liferay.commerce.account.web.internal.frontend.CommerceAccountUserRolesClayTableDataSetDisplayView" %><%@
-page import="com.liferay.commerce.account.web.internal.servlet.taglib.ui.CommerceAccountScreenNavigationConstants" %><%@
+page import="com.liferay.commerce.account.web.internal.servlet.taglib.ui.constants.CommerceAccountScreenNavigationConstants" %><%@
 page import="com.liferay.commerce.exception.NoSuchAddressException" %><%@
 page import="com.liferay.commerce.model.CommerceAddress" %><%@
 page import="com.liferay.petra.string.StringBundler" %><%@
