+++ modules/apps/commerce/commerce-address-web/src/main/java/com/liferay/commerce/address/web/internal/servlet/taglib/ui/CommerceCountryDetailsScreenNavigationEntry.java
@@ -14,6 +14,7 @@
 
 package com.liferay.commerce.address.web.internal.servlet.taglib.ui;
 
+import com.liferay.commerce.address.web.internal.servlet.taglib.ui.constants.CommerceCountryScreenNavigationConstants;
 import com.liferay.commerce.model.CommerceCountry;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
