+++ modules/apps/commerce/commerce-catalog-web/src/main/java/com/liferay/commerce/catalog/web/internal/servlet/taglib/ui/constants/CommerceCatalogScreenNavigationConstants.java
@@ -12,24 +12,23 @@
  * details.
  */
 
-package com.liferay.commerce.account.group.admin.web.internal.servlet.taglib.ui;
+package com.liferay.commerce.catalog.web.internal.servlet.taglib.ui.constants;
 
 /**
- * @author Alessio Antonio Rendina
+ * @author Gianmarco Brunialti Masera
  */
-public class CommerceAccountGroupScreenNavigationConstants {
+public class CommerceCatalogScreenNavigationConstants {
 
-	public static final String CATEGORY_KEY_COMMERCE_ACCOUNT_GROUP_DETAIL =
-		"detail";
+	public static final String CATEGORY_KEY_COMMERCE_CATALOG_ACCOUNT_GROUPS =
+		"account-groups";
 
-	public static final String ENTRY_KEY_COMMERCE_ACCOUNT_GROUP_ACCOUNTS =
-		"accounts";
+	public static final String CATEGORY_KEY_COMMERCE_CATALOG_CHANNELS =
+		"channels";
 
-	public static final String ENTRY_KEY_COMMERCE_ACCOUNT_GROUP_DETAIL =
-		"detail";
+	public static final String CATEGORY_KEY_COMMERCE_CATALOG_DETAILS =
+		"details";
 
-	public static final String
-		SCREEN_NAVIGATION_KEY_COMMERCE_ACCOUNT_GROUP_GENERAL =
-			"commerce.account.group.general";
+	public static final String SCREEN_NAVIGATION_KEY_COMMERCE_CATALOG_GENERAL =
+		"commerce.catalog.general";
 
 }
