+++ modules/apps/commerce/commerce-channel-web/src/main/java/com/liferay/commerce/channel/web/internal/servlet/taglib/ui/CommerceChannelGeneralScreenNavigationEntry.java
@@ -14,6 +14,7 @@
 
 package com.liferay.commerce.channel.web.internal.servlet.taglib.ui;
 
+import com.liferay.commerce.channel.web.internal.servlet.taglib.ui.constants.CommerceChannelScreenNavigationConstants;
 import com.liferay.commerce.product.model.CommerceChannel;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
