--- modules/apps/commerce/commerce-channel-web/src/main/java/com/liferay/commerce/channel/web/internal/servlet/taglib/ui/CommerceChannelScreenNavigationConstants.java
@@ -1,30 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.commerce.channel.web.internal.servlet.taglib.ui;
-
-/**
- * @author Marco Leo
- */
-public class CommerceChannelScreenNavigationConstants {
-
-	public static final String CATEGORY_KEY_COMMERCE_CHANNEL_GENERAL =
-		"general";
-
-	public static final String CATEGORY_KEY_COMMERCE_CHANNEL_SITE = "site";
-
-	public static final String SCREEN_NAVIGATION_KEY_COMMERCE_CHANNEL_GENERAL =
-		"commerce.channel.general";
-
-}
