+++ modules/apps/commerce/commerce-channel-web/src/main/resources/META-INF/resources/init.jsp
@@ -35,7 +35,7 @@ page import="com.liferay.commerce.channel.web.internal.frontend.CommerceChannelH
 page import="com.liferay.commerce.channel.web.internal.frontend.CommercePaymentMethodClayTable" %><%@
 page import="com.liferay.commerce.channel.web.internal.frontend.CommerceShippingMethodClayTable" %><%@
 page import="com.liferay.commerce.channel.web.internal.frontend.CommerceTaxMethodClayTable" %><%@
-page import="com.liferay.commerce.channel.web.internal.servlet.taglib.ui.CommerceChannelScreenNavigationConstants" %><%@
+page import="com.liferay.commerce.channel.web.internal.servlet.taglib.ui.constants.CommerceChannelScreenNavigationConstants" %><%@
 page import="com.liferay.commerce.constants.CommerceOrderConstants" %><%@
 page import="com.liferay.commerce.currency.model.CommerceCurrency" %><%@
 page import="com.liferay.commerce.pricing.constants.CommercePricingConstants" %><%@
